[fix/#31] 식당 id Long형으로 변경

+++ shared/feature/evaluate/src/commonMain/kotlin/com/kus/feature/evaluate/model/EvaluateRestaurant.kt
@@ -1,7 +1,7 @@
 package com.kus.feature.evaluate.model
 
 data class EvaluateRestaurant(
-    val restaurantId: Int,
+    val restaurantId: Long,
     val restaurantName: String,
     val mainTier: Int,
     val restaurantCuisine: String,
